refactor: extract chat history into AiChatHistory component
